add dtos and change restaurant componenet

+++ src/main/java/com/ecommerce/orderevent/service/RestaurantService.java
@@ -1,5 +1,8 @@
 package com.ecommerce.orderevent.service;
 
+import com.ecommerce.orderevent.dtos.RestaurantRequestDto;
+import com.ecommerce.orderevent.dtos.RestaurantResponseDto;
+import com.ecommerce.orderevent.entity.MenuItem;
 import com.ecommerce.orderevent.entity.Restaurant;
 import com.ecommerce.orderevent.exception.ResourceNotFoundException;
 import com.ecommerce.orderevent.repository.RestaurantRepository;
@@ -14,16 +17,56 @@ public class RestaurantService {
         this.restaurantRepository = restaurantRepository;
     }
 
-    public Restaurant addRestaurant(Restaurant restaurant){
-        if(restaurant.getMenuItems() != null){
-            restaurant.getMenuItems().forEach(item -> item.setRestaurant(restaurant));
+    public RestaurantResponseDto addRestaurant(RestaurantRequestDto restaurantRequestDto){
+        Restaurant restaurant = new Restaurant();
+        restaurant.setName(restaurantRequestDto.getName());
+        restaurant.setContact(restaurantRequestDto.getContact());
+        restaurant.setAddress(restaurantRequestDto.getAddress());
+
+        if (restaurantRequestDto.getMenuItems() != null) {
+            List<MenuItem> menuItems = restaurantRequestDto.getMenuItems().stream()
+                    .map(itemDto -> {
+                        MenuItem item = new MenuItem();
+                        item.setName(itemDto.getName());
+                        item.setDescription(itemDto.getDescription());
+                        item.setPrice(itemDto.getPrice());
+                        item.setRestaurant(restaurant); // set relationship
+                        return item;
+                    })
+                    .toList();
+            restaurant.setMenuItems(menuItems);
         }
-        return restaurantRepository.save(restaurant);
+        Restaurant saveRestaurant = restaurantRepository.save(restaurant);
+        return RestaurantResponseDto.fromEntity(saveRestaurant);
     }
 
-    public List<Restaurant> getAllRestaurant(){
-        return restaurantRepository.findAll();
+    public RestaurantResponseDto updateRestaurant(Long id, RestaurantRequestDto restaurantRequestDto) {
+        Restaurant existingRestaurant = restaurantRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException(RESTAURANT_NOT_FOUND+ id));
+
+        // Update allowed fields
+        existingRestaurant.setName(restaurantRequestDto.getName());
+        existingRestaurant.setAddress(restaurantRequestDto.getAddress());
+        existingRestaurant.setContact(restaurantRequestDto.getContact());
+
+        if (restaurantRequestDto.getMenuItems() != null) {
+            List<MenuItem> menuItems = restaurantRequestDto.getMenuItems().stream()
+                    .map(itemDto -> {
+                        MenuItem item = new MenuItem();
+                        item.setName(itemDto.getName());
+                        item.setDescription(itemDto.getDescription());
+                        item.setPrice(itemDto.getPrice());
+                        item.setRestaurant(existingRestaurant); // set relationship
+                        return item;
+                    })
+                    .toList();
+            existingRestaurant.setMenuItems(menuItems);
+        }
+        Restaurant restaurant = restaurantRepository.save(existingRestaurant);
+        return RestaurantResponseDto.fromEntity(restaurant);
     }
+
+    public List<Restaurant> getAllRestaurant(){ return restaurantRepository.findAll(); }
 
     public Restaurant getRestaurantById(Long id){
         return restaurantRepository.findById(id)
@@ -35,23 +78,8 @@ public class RestaurantService {
                 .orElseThrow(() -> new ResourceNotFoundException(RESTAURANT_NOT_FOUND+ id));
     }
 
-    public Restaurant updateRestaurant(Long id, Restaurant updatedRestaurant) {
-        Restaurant existingRestaurant = restaurantRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException(RESTAURANT_NOT_FOUND+ id));
-
-        // Update allowed fields
-        existingRestaurant.setName(updatedRestaurant.getName());
-        existingRestaurant.setAddress(updatedRestaurant.getAddress());
-        existingRestaurant.setContact(updatedRestaurant.getContact());
-
-        if (updatedRestaurant.getMenuItems() != null) {
-            updatedRestaurant.getMenuItems().forEach(item -> item.setRestaurant(existingRestaurant));
-            existingRestaurant.setMenuItems(updatedRestaurant.getMenuItems());
-        }
-        return restaurantRepository.save(existingRestaurant);
-    }
-
     public void deleteRestaurant(Long id){
+        if(!restaurantRepository.existsById(id)) throw new ResourceNotFoundException(RESTAURANT_NOT_FOUND+ id);
         restaurantRepository.deleteById(id);
     }
 }
